using send_track() instead of watch()/unwatch()
This handles case where downstream server is gone or down.

The watch() approach didn't work due to race conditions, where
downstream actor would be created, run and finish immediately,
even before the watch() could be set up by the spawning actor.

Instead of doing a watched_spawn(), a more useful solution was
to use send_track().  If a downstream goes away out-of-the-blue,
and attempt to send a message to a missing downstream using
send_track() sends a notification to the tracker.  Here
tracker == sender.