Resume actors immediately
Previously, we resume()'ed new actors in main todo cycle().  That had
a race when a spawner sends a message to the new actor before the next
cycle() and the current cycle() loop delivers the message before the
new actor is first resume()'ed.

Instead, invoke/resume() the new actor's function immediately
to avoid those races.